+++ provisioning/terraform/20-route53.tf
@@ -16,7 +16,7 @@ resource "aws_route53_record" "testbed" {
 
   name = "style-kit-testbed"
   type = "CNAME"
-  ttl  = "300"
+  ttl  = 300
 
   records = [
     "${aws_cloudfront_distribution.bas-style-kit-testbed.domain_name}",